rename module to clarify import
