+++ caliscope/calibration/array_initialization/estimate_paired_pose_network.py
