+++ tests/test_camera_array.py
@@ -6,7 +6,7 @@ from pathlib import Path
 import numpy as np
 
 from caliscope import __root__
-from caliscope.calibration.array_initialization.estimate_pairwise_extrinsics import estimate_paired_pose_network
+from caliscope.calibration.array_initialization.estimate_paired_pose_network import estimate_paired_pose_network
 from caliscope.calibration.array_initialization.paired_pose_network import PairedPoseNetwork
 from caliscope.cameras.camera_array import CameraArray
 from caliscope.configurator import Configurator
